requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==5.0
django-appconf==1.0.6
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-daterangepicker==1.1.0
django-bulk-sync==3.3.0
django-bulk-update==2.2.0
django-bulk-update-or-create==0.3.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-crispy-formset-modal==0.1
django-dynamic-db-router==0.3.0
django-environ==0.11.2
django-extra-views==0.14.0
django-filter==23.5
django-htmx==1.17.2
django-image-cropping==1.7
django-js-asset==2.2.0
django-menu-generator==1.1.0
django-mptt==0.16.0
django-password-expire==0.2
django-select2==8.1.2
django-session-timeout==0.1.0
django-tables2==2.7.0
django-tarview==1.0.0
django-template-partials==23.4
django-tempus-dominus==5.1.2.17
djangoajax==3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
humanize==4.9.0
idna==3.6
mssql-django==1.4
openpyxl==3.1.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sweetify==2.3.1
tablib==3.5.0
tk==0.1.0
typing_extensions==4.9.0
urllib3==2.2.1
